refactor: path var
