+++ compiler/graph/pseudo_element_compiler.py
@@ -12,7 +12,7 @@ from typing import Any, Dict
 
 import yaml
 
-from compiler.graph import adn_base_dir
+from compiler import compiler_base_dir
 
 support_list = ["logging", "qos", "null", "ratelimit", "hotel-acl"]
 
@@ -29,7 +29,7 @@ def pseudo_gen_property(element) -> Dict[str, Dict[str, Any]]:
     property = {"request": dict(), "response": dict()}
     for spec in element.spec:
         filename = spec.split("/")[-1].replace("sql", "yaml")
-        property_file = os.path.join(adn_base_dir, "elements/property", filename)
+        property_file = os.path.join(compiler_base_dir, "elements/property", filename)
         assert os.path.isfile(property_file), f"property file for {spec} not exist"
         with open(property_file, "r") as f:
             current_dict = yaml.safe_load(f)
